chore(ci): hide debugger changes in main changelog (#288)
release-please's `exclude-paths` setting makes it so that changes to those paths don't trigger a release for the given package. Unfortunately, it doesn't make it so changes aren't added to the changelog. So instead, let's hide changes scoped to `debugger`.

We can, as it turns out, always edit the changelog in the release PR, so this doesn't have to be 100% foolproof.